requirements.txt
absl-py==0.7.0
astor==0.7.1
atari-py==0.1.7
atomicwrites==1.3.0
attrs==18.2.0
box2d-py==2.3.8
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Cython==0.29.4
decorator==4.3.2
future==0.17.1
gast==0.2.2
glfw==1.7.1
grpcio==1.18.0
-e git+https://github.com/openai/gym.git@a6ac120175c28b23d498ec2ea807f8250fad4c79#egg=gym
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lockfile==0.12.2
Markdown==3.0.1
more-itertools==5.0.0
mpi4py==3.0.0
-e git+https://github.com/openai/mujoco-py.git@399a22b1384a078f8cd02d10486cff379cdf5985#egg=mujoco_py
nbformat==4.4.0
numpy==1.16.1
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
plotly==3.6.0
pluggy==0.8.1
protobuf==3.6.1
py==1.7.0
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
pytest==4.2.0
pytest-instafail==0.4.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm==4.30.0
traitlets==4.3.2
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
